move to entries module
